config: add usepkg-exclude and usepkg-include to repos.conf
Allow configuration in repos.conf of specific packages which are to
be satisfied (or not) using binary packages. These attributes appear
under the repository section and are specific to that repository,
with behaviour otherwise identical to the command line arguments of
the same name.

Where atoms specified with usepkg-exclude or usepkg-include in
repos.conf match those for the opposing command line options then
emit a warning and override the repos.conf atoms.

Signed-off-by: Jethro Donaldson <devel@jro.nz>